neu Desertlevel + prefabs
neues Desertlevel
Verknüpfungen aktualisiert
prefabs eingefügt ind Desert + Ice

+++ RollABall_CMS20_Lesson4/Assets/Scripts/PlayerController.cs
@@ -18,7 +18,7 @@ public class PlayerController : MonoBehaviour
     public                   RawImage     targetImage;
     private                  float        m_movementX, m_movementY; //input vector components
     public                   Slider       slider;
-    public                  int        lifes;
+    public                  float        lifes;
     public                   float        maxLifes;
     private                  int          m_collectablesTotalCount, m_collectablesCounter; //everything we need to count the given collectables
